Fixes a crash on sorting groupOrder and improves RenderItems sorting.

+++ lib/three3d/core/Object3D.dart
@@ -49,7 +49,7 @@ class Object3D with EventDispatcher {
   bool receiveShadow = false;
 
   bool frustumCulled = true;
-  double renderOrder = 0.0;
+  int renderOrder = 0;
 
   // List<AnimationClip> animations = [];
